Refactor filenames and directory structure
