+++ src/main/java/liquibase/ext/spanner/datatype/CharTypeSpanner.java
