+++ src/main/java/liquibase/ext/spanner/datatype/MediumIntTypeSpanner.java
