+++ src/main/java/liquibase/ext/spanner/datatype/NCharTypeSpanner.java
